refactor: move Dropdown component to shared directory for better organization
