+++ frontend/src/components/layout/Header.tsx
@@ -8,7 +8,7 @@ import { User as UserIcon, Star as StarIcon } from 'lucide-react'
 import { useAuth } from '@/context/AuthContext'
 import SearchBar from '@/components/search/SearchBar'
 import HighlightedText from '@/components/shared/HighlightedText'
-import Dropdown from '@/components/layout/Dropdown'
+import Dropdown from '@/components/shared/Dropdown'
 
 const Header = ({
   isScrollable,
